+++ Tests/EntityFrameworkCore.SqlServer.TestApplication/EntityFrameworkCore.SqlServer.TestApplication.Domain/Entities/Not Schema/ViewFolder.cs
@@ -1,0 +1,20 @@
+using System;
+using System.Collections.Generic;
+using EntityFrameworkCore.SqlServer.TestApplication.Domain.Common;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Entities.DomainEntity", Version = "1.0")]
+
+namespace EntityFrameworkCore.SqlServer.TestApplication.Domain.Entities.NotSchema
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    [DefaultIntentManaged(Mode.Fully, Targets = Targets.Properties)]
+    [DefaultIntentManaged(Mode.Fully, Targets = Targets.Methods | Targets.Constructors, Body = Mode.Ignore, AccessModifiers = AccessModifiers.Public)]
+    public class ViewFolder : IHasDomainEvent
+    {
+        public Guid Id { get; set; }
+
+        public List<DomainEvent> DomainEvents { get; set; } = new List<DomainEvent>();
+    }
+}